Move InternedBlockBuilder tests to build_interned_block/tests.rs
Keeps build_interned_block.rs focused on implementation so diff against
build_compressed_block.rs is easier to review.
